# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"preferences": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {},
							"message": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"message",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"gender": {
						"type": "string"
					},
					"birth": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"detailAddress": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"email",
					"password",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserReviewItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"storeName": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"star": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"storeName",
					"content",
					"star"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserReviewListResponse": {
				"properties": {
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/UserReviewItem"
						},
						"type": "array"
					},
					"cursor": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"reviews",
					"cursor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserReviewListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserReviewListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ReviewItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"store": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					},
					"user": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"content",
					"store",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReviewItem"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionCreateResponse": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"storeId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"point": {
						"type": "number",
						"format": "double"
					},
					"deadline": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"title",
					"description",
					"storeId",
					"missionId",
					"point",
					"deadline"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/MissionCreateResponse"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserMissionCreateResponse": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"userMissionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"missionId",
					"userMissionId",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionCreateResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserMissionCreateResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionCreateRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"missionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InProgressMissionResponse": {
				"properties": {
					"userMissionId": {
						"type": "number",
						"format": "double"
					},
					"missionId": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"point": {
						"type": "number",
						"format": "double"
					},
					"storeId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userMissionId",
					"missionId",
					"status",
					"createdAt",
					"title",
					"description",
					"point",
					"storeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InProgressMissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/InProgressMissionResponse"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true
							}
						},
						"required": [
							"cursor"
						],
						"type": "object"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_InProgressMissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/InProgressMissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionCreateResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionCreateResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionCreateRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"store_id": {
						"type": "number",
						"format": "double"
					},
					"point": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"description",
					"store_id",
					"point"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "byo-umc",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"userId": 1,
												"preferences": [
													"한식",
													"일식"
												]
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "잘못된 요청 (비밀번호 누락 등)",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "V001",
												"message": "비밀번호가 누락되었습니다."
											},
											"data": null
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "중복된 이메일 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "U001",
												"message": "이미 존재하는 이메일입니다."
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "새로운 사용자를 등록합니다. (회원가입)",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "회원가입에 필요한 사용자 정보 및 선호 카테고리",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest",
								"description": "회원가입에 필요한 사용자 정보 및 선호 카테고리"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/reviews": {
			"get": {
				"operationId": "HandleListUserReviews",
				"responses": {
					"200": {
						"description": "리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserReviewListResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"reviews": [
													{
														"id": 10,
														"storeName": "맛있는 가게",
														"content": "정말 맛있어요!",
														"star": 5
													}
												],
												"cursor": 10
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 사용자가 작성한 리뷰 목록을 조회합니다. (커서 기반 페이지네이션)",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "조회할 사용자의 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지네이션 커서 (마지막 리뷰 ID)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/stores/{storeId}/reviews": {
			"get": {
				"operationId": "HandleListStoreReviews",
				"responses": {
					"200": {
						"description": "리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"data": [
													{
														"id": 1,
														"content": "좋아요!",
														"store": {
															"id": 10,
															"name": "스타벅스"
														},
														"user": {
															"id": 5,
															"name": "홍길동"
														}
													}
												],
												"pagination": {
													"cursor": 1
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 가게의 리뷰 목록을 조회합니다. (커서 기반 페이지네이션)",
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [
					{
						"description": "조회할 가게의 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지네이션 커서 (마지막 리뷰 ID)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/stores/{storeId}/missions": {
			"get": {
				"operationId": "HandleListStoreMissions",
				"responses": {
					"200": {
						"description": "미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionListResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"data": [
													{
														"title": "일주일 3번 방문",
														"description": "3번 방문 시 포인트 지급",
														"storeId": 10,
														"missionId": 100,
														"point": 500,
														"deadline": "2026-05-21T23:28:24.854Z"
													}
												],
												"pagination": {
													"cursor": 100
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "특정 가게의 미션 목록을 조회합니다. (커서 기반 페이지네이션)",
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [
					{
						"description": "조회할 가게의 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지네이션 커서 (마지막 미션 ID)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "HandleCreateMission",
				"responses": {
					"200": {
						"description": "미션 생성 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionCreateResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"title": "신규 미션",
												"description": "미션 설명입니다.",
												"storeId": 1,
												"missionId": 10,
												"point": 1000,
												"deadline": "2026-05-21T23:28:24.858Z"
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "ST001",
												"message": "존재하지 않는 가게입니다."
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "가게에 새로운 미션을 생성합니다.",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션을 생성할 가게의 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "생성할 미션의 상세 정보",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MissionCreateRequest",
								"description": "생성할 미션의 상세 정보"
							}
						}
					}
				}
			}
		},
...
		"/users/{userId}/missions": {
			"post": {
				"operationId": "HandleCreateUserMission",
				"responses": {
					"200": {
						"description": "미션 도전 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionCreateResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"userId": 1,
												"missionId": 5,
												"userMissionId": 10,
												"status": "in_progress"
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "이미 진행 중인 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAILED",
											"error": {
												"errorCode": "M001",
												"message": "이미 진행 중인 미션입니다."
											},
											"data": null
										}
									}
								}
							}
						}
					}
				},
				"description": "유저에게 미션을 할당합니다. (미션 도전하기)",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션에 도전할 유저의 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "도전할 미션의 정보",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserMissionCreateRequest",
								"description": "도전할 미션의 정보"
							}
						}
					}
				}
			}
		},
		"/users/{userMissionId}/complete": {
			"patch": {
				"operationId": "HandleCompleteUserMission",
				"responses": {
					"200": {
						"description": "미션 완료 처리 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionCreateResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"userId": 1,
												"missionId": 5,
												"userMissionId": 10,
												"status": "completed"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "유저 미션을 완료 처리합니다.",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"description": "완료 처리할 유저 미션의 ID",
						"in": "path",
						"name": "userMissionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/in-progress": {
			"get": {
				"operationId": "HandleListMyInProgressMissions",
				"responses": {
					"200": {
						"description": "진행 중인 미션 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_InProgressMissionListResponse_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"data": {
												"data": [
													{
														"userMissionId": 10,
														"missionId": 5,
														"status": "in_progress",
														"createdAt": "2026-05-21T23:28:24.857Z",
														"title": "일주일 3번 방문하기",
														"description": "가게를 3번 방문하여 미션을 완료하세요.",
														"point": 500,
														"storeId": 1
													}
												],
												"pagination": {
													"cursor": 10
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "유저의 진행 중인 미션 목록을 조회합니다. (커서 기반 페이지네이션)",
				"tags": [
					"Missions"
				],
				"security": [],
				"parameters": [
					{
						"description": "조회할 유저의 ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "페이지네이션 커서 (마지막 유저 미션 ID)",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}